Remove GraphContainer
